rework of HNL energy calculation. See more
First you do the collimation effect. The collimation effect lets you
get Theta from theta (being careful to split the function into
restrictions where it is invertible, if the HNL is so heavy). Then the
Lorentz boost relation gives you the energy *for that Theta*. The
acceptance correction is the ratio of inverse derivatives at Theta,
|1/f'_{\pm}(Theta)|/1/f'_{SM}(Theta), where \pm indicates the
increasing (decreasing) restriction resulting from forwards or
backwards emission. The backwards emission is always at Theta >
Theta(theta = max). It is not guaranteed to exist. The likelihood of
picking + is just 1/f'_{+}(theta)/|1/f'_{-}(theta)|.