Important fix to time spent outside detector!
The `HNLVertexGenerator` placed the start of the HNL trajectory at the
`startPoint` which is now *inside* the TopVolume for geom
purposes. Explicitly save the HNL origin point in its own member. Also
breaks polarisation entirely until I can figure out how to pass the
information through the EventRecord or `SimpleHNL`.