Implemented collimation effect, see more
For multi-body LLP decays, rerun the decayer using "fudging": replace
the LLP with a massless rootino, and do a phase space decay. A new
member in Decayer tells you what that energy would be. It does not
affect the fParticles members. Then the FluxCreator can read it in and
do its calculations as usual.