Update 40Ar SF files to fix bins
This is a coercive fix to address inconsistencies between the advertised number of bins (200 on p, 400 on E) and maximum range
(2000 MeV on p, 500 MeV on E), and to bring p, E to be (presumed uniform) bin centres rather than bin edges.

The probabilities of the files have been preserved; this only changes the bins to be compatible with what
SpectralFunc assumes the format to be.