Adds SplinePostProcessor class
The SplinePostProcessor acts on a user-given set of algorithms upon spline generation.
It applies a Gaussian kernel smoothing on the spline knots' ordinates without changing the abscissae,
with user-configurable widths interpolated linearly from user-defined stops.

The intended use case is for algorithms that depend on MC throws (such as NewQELXSec) to undergo a
post-processing stage to reduce fluctuations in the splines.